SDC-7959. Call more specific JDBC set methods based on type in JDBC writer
MapD does not support setObject() in PreparedStatement. Using more specific
setType() methods where available.

Reviewed-on: https://review.streamsets.net/14050
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ jdbc-lib/src/test/java/com/streamsets/pipeline/lib/jdbc/TestJdbcMultiRowRecordWriter.java
@@ -444,7 +444,7 @@ public class TestJdbcMultiRowRecordWriter {
   private Record generateRecord() {
     Record record = RecordCreator.create();
     Map<String, Field> fields = new HashMap<>();
-    fields.put("P_ID", Field.create(++id));
+    fields.put("P_ID", Field.create((int)++id));
     fields.put("F1", Field.create(random.nextInt()));
     if (id % 2 == 0) {
       fields.put("F2", Field.create(random.nextInt()));
